generator/requirements.txt
boto3
openai
PyYAML
